requirements.txt
websockets>=12.0
httpx>=0.25.0
python-dotenv>=1.0.0
pydub>=0.25.1
